CODE REFACTOR
- Introduced MetabaseApiClient to centralize API request handling and improve code organization.
- Replaced direct API calls in MetabaseServer with calls to the new client, enhancing maintainability.
- Removed redundant error handling and caching logic from MetabaseServer, delegating responsibilities to MetabaseApiClient.
